requirements.txt
Pillow==11.1.0
rembg==2.0.62
streamlit==1.42.0
